refactor(spindrift): move the read models to the command side of the seam
The shapes commands project and views render lived in src/web/model.ts, so
fifteen modules under src/commands/ imported their own return types out of
the views' tree — the seam §21 draws, crossed in both directions.

They move to src/commands/views.ts unchanged. DeployPhase, which the module
declared as a second spelling of the deploy contract's closed union, is now
re-exported from the contract that owns it.
